Refactor: move spatial_dims code to where it's used
Removing xarray dependency in odc.geo.math. Goal is to eventually consolidate
all xarray related functionality.
